requirements.txt
audioread==3.0.1
azure-functions==1.21.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
decorator==5.1.1
idna==3.10
joblib==1.4.2
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
msgpack==1.1.0
numba==0.60.0
numpy==2.0.2
packaging==24.2
platformdirs==4.3.6
pooch==1.8.2
pycparser==2.22
requests==2.32.3
scikit-learn==1.5.2
scipy==1.14.1
soundfile==0.12.1
soxr==0.5.0.post1
threadpoolctl==3.5.0
typing_extensions==4.12.2
urllib3==2.2.3
